login amb token, arreglat comments, create i delete post (API)

+++ app/controllers/comments_controller.rb
@@ -104,7 +104,7 @@ class CommentsController < ApplicationController
 
   def api_create_comment
     @comment = Comment.new({text: params['comment']})
-    @comment.user_id = params[:user_id]
+    @comment.user_id = @api_user.id
     @comment.post_id = params[:post_id]
     @comment.tipus = 'comment'
     @comment.save
@@ -119,7 +119,7 @@ class CommentsController < ApplicationController
   def api_create_reply
     @comment = Comment.new({text: params['comment']})
     @comment.parent_id = params[:parent_id]
-    @comment.user_id = params[:user_id]
+    @comment.user_id = @api_user.id
     @comment.post_id = params[:post_id]
     @comment.tipus = 'reply'
     @comment.save
@@ -132,12 +132,16 @@ class CommentsController < ApplicationController
 
   def api_delete_comment
     @comment = Comment.find(params[:id])
+	if @comment.user_id = @api_user.id
     #@comment.destroy
-    if @comment.destroy
-      render json: @comment, status: :ok
-    else
-      render json: @comment.errors, status: :bad_request
-    end
+		if @comment.destroy
+		render json: @comment, status: :ok
+		else
+		render json: @comment.errors, status: :bad_request
+		end
+	else
+		render json: {:error => 'Unauthorized'}.to_json, :status => 401
+	end
   end
   ##end API CALLS
